Rename `peaksCurveStrength` -> `personInflateAmount`
There's a convention in this project of having high-level macro params
that control low-level variables, used as inputs in the slopes
generator.

In some cases, like with our `lineOcclusion` boolean, we just pass
a raw value through: the param is the same as the variable.

`peaksCurveStrength` was different: it shared a name across the macro
params and generator variables, but the value was different: the
`Slopes.params.js` function was doing some data-munging and then
passing it down as the same name! This is confusing, since you could
`console.log(peaksCurveStrength)` in two different steps in the same
stack and come away with different values.

This change renames the parameter from `peaksCurveStrength` to
`personInflateAmount`. This disambiguates it, as well as helps align
the parameter name to the control visualization. Note that the
generator function still receives a `peaksCurveStrength`.